backend/requirements.txt
# Flask and core dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database dependencies
PyMySQL==1.1.0
SQLAlchemy==1.4.53

# Authentication and security
bcrypt==4.1.2
python-dotenv==1.0.0

# Validation and utilities
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0

# Development dependencies
python-dateutil==2.8.2
pytest==7.4.3
pytest-flask==1.3.0